Bump to v5.8

+++ fs/cifs/transport.c
@@ -528,7 +528,7 @@ wait_for_free_credits(struct TCP_Server_Info *server, const int num_credits,
 		      const int timeout, const int flags,
 		      unsigned int *instance)
 {
-	long rc;
+	int rc;
 	int *credits;
 	int optype;
 	long int t;
